+++ src/lib/webrtc-service.ts
@@ -21,7 +21,12 @@ export interface Participant {
     name?: string,
     audio: MediaStreamTrack | null,
     video: MediaStreamTrack | null,
-    screen: MediaStreamTrack | null,
+}
+
+declare global {
+    interface RTCPeerConnection {
+        id: number;
+    }
 }
 
 export class WebRTCService {
@@ -32,7 +37,7 @@ export class WebRTCService {
             { urls: "stun:stun1.l.google.com:19302" }
         ],
     };
-    private peerConnection: Map<number, RTCPeerConnection>;
+    private peerConnection: RTCPeerConnection[];
     private socket: WebSocket | null = null
     private stream: MediaStream | null = null;
 
@@ -42,23 +47,33 @@ export class WebRTCService {
     private getMediaStreamCallback: ((id: number, stream: MediaStreamTrack | null) => void) | null = null
     private getParticipantsCallback: ((participants: Participant) => void) | null = null
 
+
     // Singleton instance
     private static instance: WebRTCService | null = null
-    public static getInstance(): WebRTCService {
-        if (!WebRTCService.instance) {
-            WebRTCService.instance = new WebRTCService();
+    public static async getInstance(roomId: string): Promise<WebRTCService | null> {
+        if (!this.instance) {
+            this.instance = new WebRTCService();
         }
-        return WebRTCService.instance
+        try {
+            if (!this.instance.socket) {
+                await this.instance.connect(roomId);
+            }
+        } catch (error) {
+            console.error("Error connecting to signaling server:", error);
+            this.instance.leave();
+            this.instance = null
+        }
+        return this.instance;
     }
 
 
     // Private constructor to prevent instantiation
     private constructor() {
-        this.peerConnection = new Map<number, RTCPeerConnection>();
+        this.peerConnection = [];
     }
 
     // Connect to signaling server
-    public async connect(roomId: string): Promise<void> {
+    private connect(roomId: string): Promise<void> {
         const serverUrl = process.env.NEXT_PUBLIC_SIGNALING_SERVER_URL || "ws://localhost:8080"
 
         return new Promise((resolve, reject) => {
@@ -113,7 +128,7 @@ export class WebRTCService {
         const videoTrack = stream.getVideoTracks()[0];
         const audioTrack = stream.getAudioTracks()[0];
 
-        for (const [, pc] of this.peerConnection) {
+        for (const pc of this.peerConnection) {
             let videoTransceiver = pc.getTransceivers().find(t => t.sender.track?.kind === 'video');
             if (!videoTransceiver) {
                 videoTransceiver = pc.addTransceiver('video', { direction: 'sendrecv' });
@@ -135,7 +150,7 @@ export class WebRTCService {
     public stopMediaStream(): void {
         if (!this.stream) return;
         this.stream.getTracks().forEach(track => track.stop());
-        for (const [, pc] of this.peerConnection) {
+        for (const pc of this.peerConnection) {
             const transceivers = pc.getTransceivers();
             transceivers.forEach(transceiver => {
                 transceiver.direction = 'recvonly';
@@ -154,130 +169,12 @@ export class WebRTCService {
             name: "Hulk",
             audio: null,
             video: null,
-            screen: null,
         })
     }
 
     public onParticipantLeave(callback: (id: number) => void): void {
         this.onParticipantLeaveCallback = callback
     }
-
-    // Start local media
-    // public async startLocalMedia(videoEnabled = true, audioEnabled = true): Promise<MediaStream> {
-    //     try {
-    //         const constraints = {
-    //             video: videoEnabled,
-    //             audio: audioEnabled,
-    //         }
-
-    //         this.localStream = await navigator.mediaDevices.getUserMedia(constraints)
-
-    //         // Add tracks to peer connection
-    //         if (this.peerConnection) {
-    //             this.localStream.getTracks().forEach((track) => {
-    //                 if (this.peerConnection && this.localStream) {
-    //                     this.peerConnection.addTrack(track, this.localStream)
-    //                 }
-    //             })
-    //         }
-
-    //         return this.localStream
-    //     } catch (error) {
-    //         console.error("Error starting local media:", error)
-    //         throw error
-    //     }
-    // }
-
-    // Toggle video
-    // public toggleVideo(enabled: boolean): void {
-    //     if (this.localStream) {
-    //         const videoTracks = this.localStream.getVideoTracks()
-    //         videoTracks.forEach((track) => {
-    //             track.enabled = enabled
-    //         })
-
-    //         // Notify other participants about the change
-    //         this.sendDataMessage({
-    //             type: "media-state-change",
-    //             videoEnabled: enabled,
-    //         })
-    //     }
-    // }
-
-    // // Toggle audio
-    // public toggleAudio(enabled: boolean): void {
-    //     if (this.localStream) {
-    //         const audioTracks = this.localStream.getAudioTracks()
-    //         audioTracks.forEach((track) => {
-    //             track.enabled = enabled
-    //         })
-
-    //         // Notify other participants about the change
-    //         this.sendDataMessage({
-    //             type: "media-state-change",
-    //             audioEnabled: enabled,
-    //         })
-    //     }
-    // }
-
-    // // Start screen sharing
-    // public async startScreenSharing(): Promise<MediaStream> {
-    //     try {
-    //         const screenStream = await navigator.mediaDevices.getDisplayMedia({
-    //             video: true,
-    //             audio: true,
-    //         })
-
-    //         // Replace video track with screen sharing track
-    //         if (this.peerConnection && this.localStream) {
-    //             const videoTrack = screenStream.getVideoTracks()[0]
-
-    //             const senders = this.peerConnection.getSenders()
-    //             const videoSender = senders.find((sender) => sender.track?.kind === "video")
-
-    //             if (videoSender) {
-    //                 videoSender.replaceTrack(videoTrack)
-    //             }
-
-    //             // Notify other participants about screen sharing
-    //             this.sendDataMessage({
-    //                 type: "screen-sharing",
-    //                 isSharing: true,
-    //             })
-    //         }
-
-    //         return screenStream
-    //     } catch (error) {
-    //         console.error("Error starting screen sharing:", error)
-    //         throw error
-    //     }
-    // }
-
-    // // Stop screen sharing
-    // public async stopScreenSharing(): Promise<void> {
-    //     try {
-    //         if (this.peerConnection && this.localStream) {
-    //             // Get original video track from local stream
-    //             const videoTrack = this.localStream.getVideoTracks()[0]
-
-    //             const senders = this.peerConnection.getSenders()
-    //             const videoSender = senders.find((sender) => sender.track?.kind === "video")
-
-    //             if (videoSender && videoTrack) {
-    //                 videoSender.replaceTrack(videoTrack)
-    //             }
-
-    //             // Notify other participants about stopping screen sharing
-    //             this.sendDataMessage({
-    //                 type: "screen-sharing",
-    //                 isSharing: false,
-    //             })
-    //         }
-    //     } catch (error) {
-    //         console.error("Error stopping screen sharing:", error)
-    //         throw error
-    //     }
-    // }
 
     // Send a message through the data channel
     // public sendDataMessage(message: Message): void {
@@ -288,42 +185,43 @@ export class WebRTCService {
     //     }
     // }
 
-    // Leave the meeting
+
+    // Leave the meeting and clean up resources
     public leave(): void {
-        // Close peer connection
         this.peerConnection.forEach((pc) => {
             pc.getTransceivers().forEach(transceiver => transceiver.stop());
             pc.close()
         })
-        this.peerConnection.clear();
+        this.peerConnection = [];
 
-        // Stop local media stream
         this.stream?.getTracks().forEach((track) => track.stop());
 
-        // Close WebSocket
         if (this.socket) {
             this.socket.close()
         }
 
-        // Clear state
-        // this.dataChannel = null
         this.stream = null
         this.socket = null
     }
 
+
     // Private methods
     private async setupPeerConnectionListeners(data: Message): Promise<RTCPeerConnection> {
-        if (this.peerConnection.has(data.from)) {
-            return this.peerConnection.get(data.from)!;
+        // find the peer connection by id from peerConnection array
+        const existingPC = this.peerConnection.find(pc => pc.id === data.from);
+        if (existingPC) {
+            return existingPC;
         }
+
         const pc = new RTCPeerConnection(this.configuration);
-        this.peerConnection.set(data.from as number, pc);
+        pc.id = data.from;
+        this.peerConnection.push(pc);
+
         this.getParticipantsCallback?.({
             id: data.from,
             name: data.name,
             audio: null,
             video: null,
-            screen: null,
         });
 
         const audioTransceiver = pc.addTransceiver('audio', { direction: this.stream ? 'sendrecv' : 'recvonly' });
@@ -343,6 +241,7 @@ export class WebRTCService {
         ]);
 
         await videoTransceiver.sender.setParameters(params);
+
         videoTransceiver.sender.replaceTrack(this.stream?.getVideoTracks()[0] || null);
         audioTransceiver.sender.replaceTrack(this.stream?.getAudioTracks()[0] || null);
 
@@ -365,6 +264,10 @@ export class WebRTCService {
             track.onended = () => {
                 console.log("Track ended")
                 this.getMediaStreamCallback?.(data.from, null);
+                pc.getTransceivers().forEach(transceiver => transceiver.stop());
+                pc.close();
+                this.peerConnection = this.peerConnection.filter(p => p !== pc);
+                this.onParticipantLeaveCallback?.(data.from)
             }
             track.onmute = () => {
                 console.log("Track muted")
@@ -385,6 +288,7 @@ export class WebRTCService {
                 })
             }
         }
+
         // Handle disconnection
         pc.onconnectionstatechange = () => {
             if (
@@ -393,9 +297,10 @@ export class WebRTCService {
                 pc.connectionState === "closed"
             ) {
                 console.log("Peer connection disconnected or failed");
-                pc?.getTransceivers().forEach(transceiver => transceiver.stop());
-                pc?.close()
-                this.peerConnection.delete(data.from);
+                pc.getTransceivers().forEach(transceiver => transceiver.stop());
+                pc.close();
+                this.peerConnection = this.peerConnection.filter(p => p !== pc);
+                this.onParticipantLeaveCallback?.(data.from)
             }
 
             if (pc.connectionState === "connected") {
@@ -495,26 +400,31 @@ export class WebRTCService {
     //     }
     // }
 
+    // Handle signaling messages
     private async handleSignalingMessage(event: MessageEvent): Promise<void> {
         try {
             const data = JSON.parse(event.data);
             switch (data.type) {
+                case WebRTCEvents.USER_JOINED:
+                    const pc = await this.setupPeerConnectionListeners(data);
+                    this.createOffer(pc, data)
+                    break
                 case WebRTCEvents.OFFER:
-                    await this.handleOffer(data)
+                    await this.handleOffer(data);
                     break
                 case WebRTCEvents.ANSWER:
-                    await this.handleAnswer(data)
+                    await this.handleAnswer(data);
                     break
                 case WebRTCEvents.ICE_CANDIDATE:
-                    await this.handleICECandidate(data)
-                    break
-                case WebRTCEvents.USER_JOINED:
-                    this.handleUserJoined(data);
+                    await this.handleICECandidate(data);
                     break
                 case WebRTCEvents.USER_LEFT:
-                    this.peerConnection.get(data.from)?.getTransceivers().forEach(transceiver => transceiver.stop());
-                    this.peerConnection.get(data.from)?.close()
-                    this.peerConnection.delete(data.from)
+                    console.log("User left")
+                    const pcToRemove = this.peerConnection.find(pc => pc.id === data.from);
+                    if (!pcToRemove) return
+                    pcToRemove.getTransceivers().forEach(transceiver => transceiver.stop());
+                    pcToRemove.close();
+                    this.peerConnection = this.peerConnection.filter(pc => pc.id !== data.from);
                     this.onParticipantLeaveCallback?.(data.from)
                     break;
                 case WebRTCEvents.DATA_MESSAGE:
@@ -536,43 +446,38 @@ export class WebRTCService {
             const answer = await pc.createAnswer()
             await pc.setLocalDescription(answer)
 
-            if (this.socket && this.socket.readyState === WebSocket.OPEN) {
-                console.log("Sending answer")
-                this.socket.send(
-                    JSON.stringify({
-                        type: WebRTCEvents.ANSWER,
-                        to: data.from,
-                        data: answer,
-                    }),
-                )
-            }
+            console.log("Sending answer")
+            this.sendMessage({
+                type: WebRTCEvents.ANSWER,
+                from: data.to,
+                to: data.from,
+                data: answer,
+            })
         } catch (error) {
             console.error("Error handling offer", error)
         }
     }
 
+
     private async handleAnswer(data: Message): Promise<void> {
         try {
             console.log("Received answer")
-            await this.peerConnection.get(data.from)?.setRemoteDescription(new RTCSessionDescription(data.data))
+            await this.peerConnection.find(p => p.id === data.from)?.setRemoteDescription(new RTCSessionDescription(data.data))
         } catch (error) {
             console.error("Error handling answer:", error)
         }
     }
 
+
     private async handleICECandidate(data: Message): Promise<void> {
-        if (!this.peerConnection.get(data.from)?.remoteDescription) return;
+        if (!this.peerConnection.find(p => p.id === data.from)?.remoteDescription) return;
         try {
-            await this.peerConnection.get(data.from)?.addIceCandidate(new RTCIceCandidate(data.data))
+            await this.peerConnection.find(p => p.id === data.from)?.addIceCandidate(new RTCIceCandidate(data.data))
         } catch (error) {
             console.error("Error handling ICE candidate:", error);
         }
     }
 
-    private async handleUserJoined(data: Message): Promise<void> {
-        const pc = await this.setupPeerConnectionListeners(data);
-        this.createOffer(pc, data)
-    }
 
     private handleDataMessage(data: any): void {
     }
@@ -592,15 +497,12 @@ export class WebRTCService {
             const offer = await pc.createOffer()
             await pc.setLocalDescription(offer)
 
-            if (this.socket && this.socket.readyState === WebSocket.OPEN) {
-                console.log("Sending offer")
-                this.sendMessage({
-                    type: WebRTCEvents.OFFER,
-                    from: -1,
-                    to: data.from,
-                    data: offer,
-                })
-            }
+            this.sendMessage({
+                type: WebRTCEvents.OFFER,
+                from: data.to,
+                to: data.from,
+                data: offer,
+            })
         } catch (error) {
             console.error("Error creating offer:", error)
         }
